version 0.5.1 i was wiriting this code while was listening to maniac by michael sembello at 6am; help me

+++ src/vm/isa/system.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
 //alone
-#include "vm/info.hpp"
+#include "vm_context.hpp"
 #include "vm/virtual_machine.hpp"
 
 namespace alone::vm::isa {
 	inline size_t halt(const context_t& ctx) {
-		ctx.flags()[flags_set::cf] = false;
+		ctx.flags()[(size_t) flags_set::cf] = false;
 		return inst_code_size;
 	}
 }
